backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
werkzeug==3.1.6